backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Deprecated==1.2.14
dj-rest-auth==4.0.1
Django==4.2.4
django-allauth==0.54.0
django-cors-headers==4.2.0
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-social-oauth2==2.1.3
django-seed==0.3.1
ecdsa==0.18.0
Faker==20.0.3
gunicorn==21.2.0
idna==3.4
jwcrypto==1.5.0
oauthlib==3.2.2
packaging==23.2
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
toposort==1.10
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wrapt==1.16.0
